requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
pydantic==1.10.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyMySQL==1.0.3
python-jose==3.3.0
alembic==1.11.1
